+++ macros/src/types/named.rs
@@ -1,6 +1,6 @@
 use proc_macro2::TokenStream;
 use quote::{quote, ToTokens};
-use syn::{spanned::Spanned, Field, FieldsNamed, GenericArgument, Generics, PathArguments, Result, Type};
+use syn::{spanned::Spanned, Field, FieldsNamed, GenericArgument, PathArguments, Result, Type};
 
 use crate::{
     attr::{FieldAttr, Inflection, Optional, StructAttr},
@@ -9,15 +9,11 @@ use crate::{
     DerivedTS,
 };
 
-pub(crate) fn named(
-    attr: &StructAttr,
-    name: &str,
-    fields: &FieldsNamed,
-    generics: &Generics,
-) -> Result<DerivedTS> {
+pub(crate) fn named(attr: &StructAttr, name: &str, fields: &FieldsNamed) -> Result<DerivedTS> {
     let mut formatted_fields = Vec::new();
     let mut flattened_fields = Vec::new();
     let mut dependencies = Dependencies::default();
+
     if let Some(tag) = &attr.tag {
         let formatted = format!("{}: \"{}\",", tag, name);
         formatted_fields.push(quote! {
@@ -32,7 +28,6 @@ pub(crate) fn named(
             &mut dependencies,
             field,
             &attr.rename_all,
-            generics,
         )?;
     }
 
@@ -48,7 +43,6 @@ pub(crate) fn named(
     };
 
     Ok(DerivedTS {
-        generics: generics.clone(),
         inline: quote!(#inline.replace(" } & { ", " ")),
         inline_flattened: Some(quote!(format!("{{ {} }}", #fields))),
         docs: attr.docs.clone(),
@@ -56,6 +50,7 @@ pub(crate) fn named(
         export: attr.export,
         export_to: attr.export_to.clone(),
         ts_name: name.to_owned(),
+        concrete: attr.concrete.clone(),
     })
 }
 
@@ -75,7 +70,6 @@ fn format_field(
     dependencies: &mut Dependencies,
     field: &Field,
     rename_all: &Option<Inflection>,
-    _generics: &Generics,
 ) -> Result<()> {
     let FieldAttr {
         type_as,
@@ -121,8 +115,12 @@ fn format_field(
     if flatten {
         match (&type_as, &type_override, &rename, inline) {
             (Some(_), _, _, _) => syn_err_spanned!(field; "`as` is not compatible with `flatten`"),
-            (_, Some(_), _, _) => syn_err_spanned!(field; "`type` is not compatible with `flatten`"),
-            (_, _, Some(_), _) => syn_err_spanned!(field; "`rename` is not compatible with `flatten`"),
+            (_, Some(_), _, _) => {
+                syn_err_spanned!(field; "`type` is not compatible with `flatten`")
+            }
+            (_, _, Some(_), _) => {
+                syn_err_spanned!(field; "`rename` is not compatible with `flatten`")
+            }
             (_, _, _, true) => syn_err_spanned!(field; "`inline` is not compatible with `flatten`"),
             _ => {}
         }
@@ -178,7 +176,9 @@ fn extract_option_argument(ty: &Type) -> Result<&Type> {
                         other => syn_err!(other.span(); "`Option` argument must be a type"),
                     }
                 }
-                other => syn_err!(other.span(); "`Option` type must have a single generic argument"),
+                other => {
+                    syn_err!(other.span(); "`Option` type must have a single generic argument")
+                }
             }
         }
         other => syn_err!(other.span(); "`optional` can only be used on an Option<T> type"),